# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beampipe",
    "description": "Beampipe Core\n\n## Getting started\n\n1. **Authenticate**: `POST /api/v1/login` with the admin **username or email** and password from setup. Copy `access_token` and click **Authorize** (Bearer token).\n2. **List projects**: `GET /api/v1/projects` shows registered project modules.\n3. **Register sources**: `POST /api/v1/sources` (returns 201 or 200 if already registered).\n4. **Run discovery**: `POST /api/v1/sources/discover` marks sources for async archive polling.\n5. **Create execution**: `POST /api/v1/executions/prepare` (preview) then `POST /api/v1/executions`.\n6. **Execute**: `POST /api/v1/executions/{id}/execute` enqueues staging/submit on the worker queue.\n7. **Poll**: `GET /api/v1/executions/{id}/status` or `.../ledger-snapshot` for operator view.\n\n### Common errors\n\n- **401**: missing or expired token. Log in again.\n- **400**: unknown `project_module` (check `GET /api/v1/projects`).\n- **404**: source, execution, or deployment profile not found.\n- **503**: Redis worker queue unavailable.\n\nHealth: `GET /api/v1/health` (liveness) \u00b7 `GET /api/v1/ready` (deps).\n\nLocal-only UI: [`/sources`](/sources) for a minimal browser view.",
    "version": "0.1.0",
    "contact": {
      "name": "admin",
      "email": "admin@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List sources",
        "description": "List registered sources with optional filters and pagination.",
        "operationId": "list_sources_api_v1_sources_get",
        "parameters": [
          {
            "name": "project_module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project module identifier",
              "title": "Project Module"
            },
            "description": "Filter by project module identifier"
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by enabled flag",
              "title": "Enabled"
            },
            "description": "Filter by enabled flag"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Items Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SourceSortField",
              "description": "Column to sort by",
              "default": "created_at"
            },
            "description": "Column to sort by"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort direction",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated source registry entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse_SourceRegistryRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Register source",
        "description": "Register a source in the registry. Returns 200 when the source already exists.",
        "operationId": "register_source_api_v1_sources_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRegistryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New source created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryRead"
                }
              }
            }
          },
          "200": {
            "description": "Source already registered (idempotent)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. unknown project module)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/bulk": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Bulk register sources",
        "description": "Register many sources in one request. Existing sources are returned in `existing` without error.",
        "operationId": "bulk_register_sources_api_v1_sources_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRegistryBulkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryBulkCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. unknown project module)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sources/discover": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Trigger discovery",
        "description": "Mark sources for rediscovery. Discovery runs asynchronously via the background scheduler.",
        "operationId": "trigger_discovery_api_v1_sources_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverTriggerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. unknown project module)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/sources/{source_id}/executions": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List source executions",
        "description": "List batch executions that included this registry source.",
        "operationId": "list_source_executions_api_v1_sources__source_id__executions_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Page size (max 50)",
              "default": 10,
              "title": "Items Per Page"
            },
            "description": "Page size (max 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse_SourceLinkedExecutionItem_"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get source",
        "description": "Return a single source registry entry by UUID.",
        "operationId": "get_source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryRead"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sources"
        ],
        "summary": "Update source",
        "description": "Update enabled flag and/or stale-after hours for a source.",
        "operationId": "update_source_api_v1_sources__source_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRegistryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sources"
        ],
        "summary": "Delete source",
        "description": "Permanently remove a source from the registry.",
        "operationId": "delete_source_api_v1_sources__source_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/metadata": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get source metadata",
        "description": "Return archive metadata rows linked to this source.",
        "operationId": "get_source_metadata_api_v1_sources__source_id__metadata_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedListResponse_SourceRegistryRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SourceRegistryRead"
            },
            "type": "array",
            "title": "Data"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "items_per_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items Per Page"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total_count",
          "has_more"
        ],
        "title": "PaginatedListResponse[SourceRegistryRead]"
      },
...
      "SourceSortField": {
        "type": "string",
        "enum": [
          "created_at",
          "updated_at",
          "source_identifier",
          "last_checked_at"
        ],
        "title": "SourceSortField"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login"
          }
        }
      }
    }
  },
...
  "servers": [
    {
      "url": "/",
...
    }
  ],
...
}